stm32mp1: low_power: manage self refresh in CSTOP for LPDDR2
Self Refresh (SR) is a significant power saving opportunity for SDRAM.
During SR, DLL's in DDRPHYC can be bypassed,and the DDRCTRL clocks can be
stopped.
Automatic SR (ASR) triggers on when there are no read/write for a
predefined time. This fits with the need in running mode.
Software SR (SSR) is controlled by software operation.

Switch to software self refresh when entering in CSTOP, and go back to
automatic self refresh when exiting from CSTOP.

Change-Id: I3ce91d2124b63491cc88eb39de4c643e59bd135c
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/152785
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>