pmic: stm32mp_pmic: only program voltage if not in DT range
When setting boot-on regulator, the voltage register is only re-programmed
if the read value in PMIC is not in the range provided in DT.
The dt_pmic_get_regulator_voltage() function is updated to read both
min and max voltage values.
The macro CMD_GET_VOLTAGE is renamed to CMD_GET_MIN_VOLTAGE to reflect
what it really does.

Change-Id: I101190de63bb0434990dcf9e62be322375aa68a4
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/159074
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>