stm32mp1: correctly check return from STANDBY
When returning from STANDBY state, only checking the boot action status
is not enough. We also have to check the non-secure return address is
correct, and the saved context is OK in backup RAM.
This is required for STANDBY with DDR off, which can be used when a user
issues a shutdown.

Change-Id: I4b1d5159cd7a524eb893229ea9bcf628746de0ab
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/183406
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>