stm32mp1: remove useless DDR cache flush
The flush of the BL33 image is already handled in load_image().
For ED1/EV1 boards, this extra flush was also doing nothing, as for
those boards image_base + image_max_size = 0x100000000, which overflows
an u32, and 0 was used.
This flush is now replaced with a correct invalidation of the cache
directly in io_stm32image.

Change-Id: Ibe301e017669b10d8c321abef27c674c914e43b6
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/183678
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>